+++ src/main/java/com/atlauncher/workers/InstanceInstaller.java
@@ -1049,27 +1049,24 @@ public class InstanceInstaller extends SwingWorker<Boolean, Void> implements Net
                 FileUtils.deleteDirectory(this.root.resolve("config"));
             }
 
-            if (((instance != null && instance.getMinecraftVersion().equals(this.minecraftVersion.id))
-                    || (instanceV2 != null && instanceV2.id.equals(this.minecraftVersion.id)))
-                    && (instanceV2 != null ? instanceV2.hasCustomMods() : instance.hasCustomMods())) {
-                if (isReinstall) {
-                    if (Files.isDirectory(this.root.resolve("mods"))) {
-                        Utils.deleteWithFilter(this.root.resolve("mods").toFile(),
-                                (instanceV2 != null ? instanceV2.getCustomMods(com.atlauncher.data.Type.mods)
-                                        : instance.getCustomMods(com.atlauncher.data.Type.mods)));
-                    }
+            if (isReinstall) {
+                if (Files.isDirectory(this.root.resolve("mods"))) {
+                    Utils.deleteWithFilter(this.root.resolve("mods").toFile(),
+                            (instanceV2 != null ? instanceV2.getPackMods(com.atlauncher.data.Type.mods)
+                                    : instance.getPackMods(com.atlauncher.data.Type.mods)), true);
+                }
 
-                    if (Files.isDirectory(this.root.resolve("coremods"))) {
-                        Utils.deleteWithFilter(this.root.resolve("coremods").toFile(),
-                                (instanceV2 != null ? instanceV2.getCustomMods(com.atlauncher.data.Type.coremods)
-                                        : instance.getCustomMods(com.atlauncher.data.Type.coremods)));
-                    }
+                if (Files.isDirectory(this.root.resolve("coremods"))) {
+                    Utils.deleteWithFilter(this.root.resolve("coremods").toFile(),
+                            (instanceV2 != null ? instanceV2.getPackMods(com.atlauncher.data.Type.coremods)
+                                    : instance.getPackMods(com.atlauncher.data.Type.coremods)),
+                            true);
+                }
 
-                    if (Files.isDirectory(this.root.resolve("jarmods"))) {
-                        Utils.deleteWithFilter(this.root.resolve("jarmods").toFile(),
-                                (instanceV2 != null ? instanceV2.getCustomMods(com.atlauncher.data.Type.jar)
-                                        : instance.getCustomMods(com.atlauncher.data.Type.jar)));
-                    }
+                if (Files.isDirectory(this.root.resolve("jarmods"))) {
+                    Utils.deleteWithFilter(this.root.resolve("jarmods").toFile(),
+                            (instanceV2 != null ? instanceV2.getPackMods(com.atlauncher.data.Type.jar)
+                                    : instance.getPackMods(com.atlauncher.data.Type.jar)), true);
                 }
             } else {
                 FileUtils.deleteDirectory(this.root.resolve("mods"));