+++ CHANGELOG.md
@@ -23,3 +23,4 @@
 - Add in ability to update curse mods from edit mods dialog
 - Add in Servers button to packs and instances to find servers to play on
 - Add ability to drag and drop mods into the edit mods dialog
+- Only delete mods that the pack itself installed on reinstall/update